+++ .gitignore
@@ -1,3 +1,4 @@
 __pycache__
 /search/__pycache__/
 .vscode
+*.pyc
