+++ admin.py
@@ -1,0 +1,26 @@
+from flask import redirect, request, url_for
+from flask_admin import Admin, AdminIndexView
+from flask_admin.contrib.sqla import ModelView
+from flask_security import current_user
+
+from app import app, db
+from models import Inquiry, Role, User
+
+
+class AdminMixin:
+    def is_accessible(self):
+        return current_user.has_role('admin')
+
+    def inaccessible_callback(self, name,  **kwargs):
+        return redirect(url_for('security.login', next=request.url))
+
+class AdminView(AdminMixin, ModelView):
+    pass
+
+class HomeAdminView(AdminMixin, AdminIndexView):
+    pass
+
+admin = Admin(app, 'FlaskApp', url='/', index_view=HomeAdminView(name='Home'))
+admin.add_view(AdminView(User, db.session))
+admin.add_view(AdminView(Role, db.session))
+admin.add_view(AdminView(Inquiry, db.session))
