+++ app.py
@@ -1,67 +1,19 @@
 from flask import Flask
-from config import Configuration
-from flask_sqlalchemy import SQLAlchemy
-
 from flask_migrate import Migrate, MigrateCommand
 from flask_script import Manager
+from flask_sqlalchemy import SQLAlchemy
 
-from flask_admin import Admin
-from flask_admin import AdminIndexView
-from flask_admin.contrib.sqla import ModelView
+from config import Configuration
 
-from flask_security import SQLAlchemyUserDatastore
-from flask_security import Security
-
-from flask import redirect, url_for, request
-from flask_security import current_user
 
 app = Flask(__name__)
 app.config.from_object(Configuration)
+
+from heroku_config import HerokuConfig
+''' heroku_config импортирует app '''
 
 db = SQLAlchemy(app)
 
 migrate = Migrate(app, db)
 manager = Manager(app)
 manager.add_command('db', MigrateCommand)
-
-### Admin ###
-from models import *
-
-class AdminMixin:
-    def is_accessible(self):
-        return current_user.has_role('admin')
-
-    def inaccessible_callback(self, name,  **kwargs):
-        return redirect(url_for('security.login', next=request.url))
-
-class AdminView(AdminMixin, ModelView):
-    pass
-
-class HomeAdminView(AdminMixin, AdminIndexView):
-    pass
-
-admin = Admin(app, 'FlaskApp', url='/', index_view=HomeAdminView(name='Home'))
-admin.add_view(AdminView(User, db.session))
-admin.add_view(AdminView(Role, db.session))
-admin.add_view(AdminView(Inquiry, db.session))
-
-### Flask security ###
-
-user_datastore = SQLAlchemyUserDatastore(db, User, Role)
-security = Security(app, user_datastore)
-
-### Heroku ### 
-
-class HerokuConfig(Configuration):
-    @classmethod
-    def init_app(cls, app):
-        ProductionConfig.init_app(app)
-
-    import logging
-    from logging import StreamHandler
-    
-    file_handler = StreamHandler()
-    file_handler.setLevel(logging.WARNING)
-    app.logger.addHandler(file_handler)
-
-
