+++ main.py
@@ -1,9 +1,9 @@
 #!/usr/bin/python3
 # -*- coding: utf-8 -*-
 
-from app import app, db
-import view
 
+import view
+from app import app, db
 from search.blueprint import search
 
 app.register_blueprint(search, url_prefix='/search')
